kpis/requirements.txt
# Core data processing
pandas>=2.0
python-dateutil>=2.8
pytz>=2023.3

# API clients
requests>=2.31
jira>=3.5
PyGithub>=2.1

# Configuration
pyyaml>=6.0
